pyproject.toml
[project]
name = "snipster"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Ben Griffith", email = "bengriffith@outlook.com" }
]
requires-python = ">=3.13"
dependencies = [
    "fastapi[all]>=0.115.12",
    "psycopg2>=2.9.10",
    "python-decouple>=3.8",
    "sqlmodel>=0.0.24",
    "typer>=0.16.0",
]

[project.scripts]
snipster = "snipster:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "psycopg2-binary>=2.9.10",
    "pytest-cov>=6.1.1",
]
